### 2025-04-07 #7911 @RMeli
Hi @awvwgk 👋

Would you be interested in reviewing this submission for JOSS (Journal of Open Source Software)?

You can find more information about the submission at the top of this Github issue #7911.

Thank you in advance!
