### 2025-04-07 #7911 @AlexBuccheri
Hi @RMeli I can review this in principle, however Octopus is running a TDDFT school from today until the 17th, then I'll be off over the Easter weekend. If there's not a big time pressure, and this can wait until the end of April, I'm happy to review.
