### 2025-04-07 #7911 @awvwgk
Hi Rocco @RMeli, I am happy to review this. It's great to see MOPAC around here.
