### 2025-04-07 #8025 @editorialbot
[REVIEW]: MOPAC: An open-source semiempirical molecular orbital program
**Submitting author:** @godotalgorithm (<a href="http://orcid.org/0000-0003-3701-1830">Jonathan Moussa</a>)
**Repository:** https://github.com/openmopac/mopac
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v23.2.3
