### 2025-04-14 #8025 @berquist
- Apache 2.0
  - no original data
  - no original results

  - ~~http://openmopac.net/manual/ could be made more prominent in the top-level README~~
  - http://openmopac.net/manual/examples.html
  - Eventually https://openmopac.github.io/examples.html
  - ~~Should be a simple example in the repository that is separate from the tests, or point out test cases that are sufficiently general~~ done
  - ~~Nothing for https://github.com/openmopac/mopactools~~ not the focus of this review
  - Original web manual is detailed but needs migration to https://github.com/openmopac/openmopac.github.io
  - `README.md` in each source code subdirectory documenting its high-level purpose
  - CTest works with no additional configuration
  - https://github.com/openmopac/mopac/blob/main/.github/workflows/CI.yaml
  - ~~Nothing for https://github.com/openmopac/mopactools~~ done in https://github.com/openmopac/mopactools/commit/81f5861af6f3edc3d4115e41a154dcf11a869995
  - https://github.com/openmopac/mopac/blob/main/CONTRIBUTING.rst

  - Text on the importance of MOPAC no longer being closed?
  - Could state that MOPAC is _the_ semiempirical quantum chemistry program
