### 2025-04-14 #8025 @berquist
> FYI, I created a Spack package for `mopac`. It is not complete (MDI dependency is missing, for example), and I might have missed things, but it's a good starting point for testing at least for me.
>
> The following works for me, and all tests pass:
>
> ```
> spack install -v --test=root mopac@23.1.2 +tests +openmp +cmalloc
> ```
>
> [spack/spack#49949](https://github.com/spack/spack/pull/49949)

Hah, I created one on 2024-04-07 and never made a PR for it.  I'll definitely take a look.
