### 2025-05-05 openmopac/mopac#264 @awvwgk
Website at http://openmopac.net/ does only mention MOPAC2016
The linked website for MOPAC does not reflect the current state of the project, mainly it mentions MOPAC2016 and links to a reseller webpage for packages which include MOPAC.

Only the second download link mentions that MOPAC is available as open source project.

JOSS review: https://github.com/openjournals/joss-reviews/issues/8025
