### 2025-05-05 openmopac/mopac#265 @awvwgk
Documentation for installation inconsistent
I checked several places for finding the installation instructions:

* https://openmopac.github.io/downloads.html
  * only placeholder, no real instructions
* http://openmopac.net/manual/index.html
  * only reference to MOPAC2016 and precompiled binaries
* http://openmopac.net/Download_MOPAC_Executable_Step2.html
  * provides reference to conda-forge and links to installers
* https://github.com/openmopac/mopac?tab=readme-ov-file#installation
  * mentions installers in GH release or building from source

It would be great to have one place for the installation instructions to point users to, or at least have the installation instructions in all places consistent.

JOSS review: https://github.com/openjournals/joss-reviews/issues/8025
