### 2025-05-05 openmopac/mopac#266 @awvwgk
Contributing guidelines not referenced in landing page
Contributing guide lines (https://github.com/openmopac/mopac/blob/main/CONTRIBUTING.rst) are currently not referenced from the repository README or the documentation as far as I can see. It would be great to include a statement pointing potential contributors to the document.

JOSS review: https://github.com/openjournals/joss-reviews/issues/8025
