### 2025-05-05 openmopac/mopac#267 @awvwgk
Comparison with other semiempirical programs
Maybe it could be useful to compare MOPAC with other semiempirical programs, like Sparrow with also implements PM7.

- https://scine.ethz.ch/static/download/documentation/sparrow/v3.0.0/py/index.html
