### 2025-05-05 #8025 @awvwgk
I had an initial look through the paper and documentation. One thing which was very confusing for me is that there are multiple webpages describing different states of the same (?) project.

* https://github.com/openmopac/mopac/issues/264
* https://github.com/openmopac/mopac/issues/265
* https://github.com/openmopac/mopac/issues/266
* https://github.com/openmopac/mopac/issues/267
