### 2025-05-05 openmopac/mopac#268 @awvwgk
Full example for running MOPAC library from C (or Fortran)
The README mentions access to an API, it would be great to have a full example on how to use the API from C and/or Fortran as part of the documentation and referenced from the README.

- https://github.com/openmopac/mopac?tab=readme-ov-file#run-from-library

JOSS review: https://github.com/openjournals/joss-reviews/issues/8025
