### 2025-05-08 openmopac/mopac#264 @awvwgk
I think for the JOSS paper of Mopac having a single definite documentation source is necessary, even if the material is just copied over from the Mopac webpage.
