### 2025-05-08 openmopac/mopac#264 @godotalgorithm
That's fair, it is overdue at this point. I will soon make the website migration my top priority, so as not to hold up the JOSS review for too long.

To be clear, the documentation does not have to reside in the software repository associated with the JOSS paper, correct?
