### 2025-05-09 openmopac/mopac#264 @awvwgk
I don't think there is a requirement about the source of the documentation
