### 2025-05-14 #8025 @awvwgk
From my side, it seems like there is some work regarding the documentation left to be done, also for the paper I am missing a comparison with other semi-empirical packages supporting similar functionality as MOPAC.
