+++ src/solvation/cosmo.F90
@@ -187,28 +187,6 @@ subroutine ansude (ra, rb, d, rs, ara, arb, aar, abr, arad, arbd, rinc)
     arad = pi * ra * (sad*za + sa*zad + 2.d0*ra*cad)
     arbd = pi * rb * (sbd*zb + sb*zbd + 2.d0*rb*cbd)
 end subroutine ansude
-subroutine ciint (c34, pq34)
-    use molkst_C, only : lm61
-    use cosmo_C, only : nps, cmat
-    implicit none
-    double precision, dimension (lm61), intent (in) :: c34
-    double precision, dimension (lm61), intent (inout) :: pq34
-    integer :: i, i0, j
-    i0 = 0
-    do i = 1, lm61
-      pq34(i) = 0.d0
-    end do
-    if (nps < 0) return
-    do i = 1, lm61
-      do j = 1, i - 1
-        i0 = i0 + 1
-        pq34(j) = pq34(j) + cmat(i0) * c34(i)
-        pq34(i) = pq34(i) + cmat(i0) * c34(j)
-      end do
-      i0 = i0 + 1
-      pq34(i) = pq34(i) + cmat(i0) * c34(i)
-    end do
-end subroutine ciint
 subroutine coscav
    !***********************************************************************
    !
@@ -1749,7 +1727,7 @@ end subroutine surclo
 subroutine cosini(l_print)
     use cosmo_C, only : n0, ioldcv, fnsq, nps, rsolv, nspa, disex2, &
     dirsm, dirvec, srad, ipiden, gden, idenat, qdenet, amat, &
-    cmat, lenabc, arat, sude, isude, bh, qden, nar_csm, nsetf, phinet, &
+    lenabc, arat, sude, isude, bh, qden, nar_csm, nsetf, phinet, &
     qscnet, bmat, nset, xsp, abcmat, iatsp, nn, qscat, cosurf, nppa, &
     ffact, fepsi, nden
     use common_arrays_C, only : nat, nfirst, nlast
@@ -1799,7 +1777,6 @@ subroutine cosini(l_print)
     if (allocated(qscnet)) deallocate (qscnet)
     if (allocated(abcmat)) deallocate (abcmat)
     if (allocated(bmat))   deallocate (bmat)
-    if (allocated(cmat))   deallocate (cmat)
     if (allocated(qscat))  deallocate (qscat)
     if (allocated(srad))   deallocate (srad)
     if (allocated(iatsp))  deallocate (iatsp)
@@ -1838,8 +1815,6 @@ subroutine cosini(l_print)
       j = j + i
       allocate(amat((lenabc*(lenabc + 1))/2), stat = i)
       j = j + i
-      allocate(cmat((lm61*(lm61 + 1))/2), stat = i)
-      j = j + i
       if (j /= 0) then
         if (l_print) then
           write(line, '(a, i5, a)')"Data set '"//trim(jobnam)//"' exists, "
@@ -1855,7 +1830,6 @@ subroutine cosini(l_print)
         call memory_error("COSINI (2) in Cosmo")
         return
       end if
-      cmat = 0.d0
     end if
     call extvdw (usevdw, rvdw)
     if (moperr) return
