### 2025-05-21 41cc0d5
Remove remaining unused COSMO arrays
