+++ src/setup_mopac_arrays.F90
@@ -53,9 +53,9 @@
   use drc_C, only: vref, vref0, allxyz, allvel, xyz3, vel3, allgeo, geo3, parref
 !
   use cosmo_C, only : ipiden, idenat, gden, qdenet, &
-  phinet, qscnet, qscat, bmat, abcmat, srad, iatsp, &
-  nn, cosurf, xsp, nset, bh, qden, nsetf, isude, sude, &
-  arat, amat, cmat
+  phinet, qscnet, qscat, bmat, srad, iatsp, &
+  nn, cosurf, nset, qden, nsetf, isude, sude, &
+  arat, amat
 !
   use esp_C, only : cen, ex, iam, cc, &
       & cespm, espi, potpt, es, rnai, rnai2, pf0, pf1, pf2, &
@@ -194,7 +194,6 @@
     if (allocated(f)) call fock2(f, f, f, w, w, w, numat, nfirst, nlast, mode)
     if (allocated(a2))         deallocate(a2, stat = i)
     if (allocated(ab))         deallocate(ab, stat = i)
-    if (allocated(abcmat))     deallocate (abcmat, stat = i)
     if (allocated(aicorr))     deallocate (aicorr, stat = i)
     if (allocated(aidref))     deallocate(aidref, stat = i)
     if (allocated(al))         deallocate(al, stat = i)
@@ -206,7 +205,6 @@
     if (allocated(atmass))     deallocate (atmass, stat = i)
     if (allocated(b))          deallocate(b, stat = i)
     if (allocated(b_esp))      deallocate(b_esp, stat = i)
-    if (allocated(bh))         deallocate (bh, stat = i)
     if (allocated(bmat))       deallocate (bmat, stat = i)
     if (allocated(c))          deallocate (c, stat = i)
     if (allocated(cb))         deallocate (cb, stat = i)
@@ -218,7 +216,6 @@
     if (allocated(cespm))      deallocate(cespm, stat = i)
     if (allocated(cespm2))     deallocate(cespm2, stat = i)
     if (allocated(cespml))     deallocate(cespml, stat = i)
-    if (allocated(cmat))       deallocate (cmat, stat = i)
     if (allocated(cnorml))     deallocate(cnorml, stat = i)
     if (allocated(co))         deallocate(co, stat = i)
     if (allocated(coord))      deallocate (coord, stat = i)
@@ -363,7 +360,6 @@
     if (allocated(work2))      deallocate(work2, stat = i)
     if (allocated(xparam))     deallocate (xparam, stat = i)
     if (allocated(xparef))     deallocate (xparef, stat = i)
-    if (allocated(xsp))        deallocate (xsp, stat = i)
     if (allocated(xyz3))       deallocate (xyz3, stat = i)
     if (allocated(occa))       deallocate(occa, stat = i)
     if (allocated(microa))     deallocate(microa, stat = i)
