### 2025-05-21 645f158
Clarify COSMO authorship
