+++ src/solvation/README.md
@@ -1,4 +1,5 @@
 # COSMO solvation model
 
-This directory contains MOPAC's implementation of the COSMO solvation model,
-which was developed and written by Andreas Klamt.
+This directory contains MOPAC's two implementations of the COSMO solvation model.
+The original, refrence implementation was developed and written by Andreas Klamt.
+The fast implementation, which includes an implementation of the fast multipole method, was written by Andrey Bliznyuk.
