### 2025-05-22 0d345ee
Install OpenBLAS through Ubuntu package manager
