### 2025-05-23 openmopac/mopac#267 @godotalgorithm
I'm inferring from your comments in the JOSS review that this is a request about the JOSS paper and not the MOPAC documentation?

I'd like to keep the JOSS paper short, but I can cite a [perspective paper](https://doi.org/10.48550/arXiv.2505.13424) that I just released about semiempirical modeling for broader context, and I can provide a short list of popular and/or notable programs that also implement some PMx models.

Very few programs have PM7 implementations (I think Gaussian is the only other program besides MOPAC) because its literature specification and implementation in MOPAC have some deviations that made independent re-implementation difficult. I've been collecting some of them as GitHub Issues (e.g. #240 and #258), although I might move them to a webpage once the old website is migrated to GitHub Pages.
