### 2025-05-23 762a84c
Add more method and software references for semiempirical models
