+++ paper.bib
@@ -1,12 +1,92 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% https://bibdesk.sourceforge.io/
 
-%% Created for Jonathan Moussa at 2025-02-22 00:58:58 -0500 
+%% Created for Jonathan Moussa at 2025-05-23 12:55:02 -0400 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
 
 
+
+@misc{Gaussian,
+	address = {Wallingford, CT %! Gaussian 16},
+	author = {Frisch, M. J. and Trucks, G. W. and Schlegel, H. B. and Scuseria, G. E. and Robb, M. A. and Cheeseman, J. R. and Scalmani, G. and Barone, V. and Petersson, G. A. and Nakatsuji, H. and Li, X. and Caricato, M. and Marenich, A. V. and Bloino, J. and Janesko, B. G. and Gomperts, R. and Mennucci, B. and Hratchian, H. P. and Ortiz, J. V. and Izmaylov, A. F. and Sonnenberg, J. L. and Williams and Ding, F. and Lipparini, F. and Egidi, F. and Goings, J. and Peng, B. and Petrone, A. and Henderson, T. and Ranasinghe, D. and Zakrzewski, V. G. and Gao, J. and Rega, N. and Zheng, G. and Liang, W. and Hada, M. and Ehara, M. and Toyota, K. and Fukuda, R. and Hasegawa, J. and Ishida, M. and Nakajima, T. and Honda, Y. and Kitao, O. and Nakai, H. and Vreven, T. and Throssell, K. and Montgomery Jr., J. A. and Peralta, J. E. and Ogliaro, F. and Bearpark, M. J. and Heyd, J. J. and Brothers, E. N. and Kudin, K. N. and Staroverov, V. N. and Keith, T. A. and Kobayashi, R. and Normand, J. and Raghavachari, K. and Rendell, A. P. and Burant, J. C. and Iyengar, S. S. and Tomasi, J. and Cossi, M. and Millam, J. M. and Klene, M. and Adamo, C. and Cammi, R. and Ochterski, J. W. and Martin, R. L. and Morokuma, K. and Farkas, O. and Foresman, J. B. and Fox, D. J.},
+	date = {2016},
+	date-added = {2025-05-23 12:54:38 -0400},
+	date-modified = {2025-05-23 12:54:57 -0400},
+	title = {Gaussian 16 Rev. C.01},
+	year = {2016}}
+
+@misc{SQM_perspective,
+	archiveprefix = {arXiv},
+	author = {Jonathan E. Moussa},
+	date-added = {2025-05-23 12:46:55 -0400},
+	date-modified = {2025-05-23 12:47:11 -0400},
+	eprint = {2505.13424},
+	primaryclass = {physics.chem-ph},
+	title = {The Enduring Relevance of Semiempirical Quantum Mechanics},
+	url = {https://arxiv.org/abs/2505.13424},
+	year = {2025},
+	bdsk-url-1 = {https://arxiv.org/abs/2505.13424}}
+
+@article{ULYSSES,
+	annote = {doi: 10.1021/acs.jcim.2c00757},
+	author = {Menezes, Filipe and Popowicz, Grzegorz M.},
+	date = {2022/08/22},
+	date-added = {2025-05-23 12:46:13 -0400},
+	date-modified = {2025-05-23 12:46:29 -0400},
+	doi = {10.1021/acs.jcim.2c00757},
+	isbn = {1549-9596},
+	journal = {Journal of Chemical Information and Modeling},
+	journal1 = {Journal of Chemical Information and Modeling},
+	journal2 = {J. Chem. Inf. Model.},
+	month = {08},
[…]
