### 2025-05-23 openmopac/mopac#267 @godotalgorithm
I just made these paper revisions in 7131657. Tell me if I need to do something else as part of a JOSS paper revision or if this doesn't resolve your Issue.
