### 2025-06-03 #8025 @godotalgorithm
@awvwgk raised concerns about the state of MOPAC's online documentation, with the [old website](http://openmopac.net) not fully reflecting MOPAC's open-source status because it is cumbersome to update and a visible, but clearly incomplete [new website](https://openmopac.github.io). I'm presently trying to resolve this by migrating the contents of the old website to GitHub Pages hosting, so that I can more easily create and deploy new top-level/landing pages to the website. Once this is done, I will deactivate the previous new website and redirect the old domain name to the GitHub-host website. The remaining issues about MOPAC will be easy to resolve after the web migration is complete, and I've already revised the paper itself following reviewer suggestions.

I just started the web migration last week, and it will probably take a few weeks before it is ready. I've already tried to move the website to other hosting in a quick and minimal way before, but there are a variety of technical problems that prevent this from being a simple job (e.g. the raw contents are too large to fit into one GitHub repo, all links and shortcuts need to be rebuilt because they are case insensitive, which is a hidden incompatibility between Windows and Linux web servers).
