### 2025-06-18 openmopac/mopac#272 @berquist
Documentation for typical example
This is for the [JOSS review](https://github.com/openjournals/joss-reviews/issues/8025).

There are many example inputs on the old website and many tests (that may or many not be good exemplars), but something simple could be added to the top-level README, or an `example.{mop,out}` in the top-level directory could be pointed to.  2016 used to come with an example input (and output? can't remember) when unzipping the archive.
