### 2025-06-18 #8025 @berquist
Most of my concerns will be addressed once the website documentation makes progress, but I had a question about the Python repository.  If it's mentioned in the paper, it should be considered a part of the review, correct?
