### 2025-06-19 openmopac/mopac#272 @godotalgorithm
As I'm dealing with the broader documentation and website migration issues, I plan to add an `examples` directory to the main repo and distribution with a set of basic example inputs (and a subdirectory containing what their successful outputs should look like). A few years ago, I did a survey of the literature to see how people use MOPAC in practice, and typical usage is rather narrow. I'll probably take a second pass at this to dig for less-common usage and prepare examples to cover the distribution of typical use, as well as recommended usage of notable features such as MOZYME (which unfortunately doesn't see much use) and the new API.
