### 2025-07-04 #8025 @godotalgorithm
> Hi [@godotalgorithm](https://github.com/godotalgorithm), how are things progressing? Please let me know if there is anything I can help with.

Unfortunately, I've been sidetracked over the last several weeks by finishing up [another paper](https://arxiv.org/abs/2503.14970). I'll be on the job market again later this year, and I'm trying to wrap up multiple projects at my current job, which has included submitting 2 single-author papers so far this year in addition to this JOSS paper. I will soon return to the MOPAC website migration as my top work priority, which is the main holdup of this peer review process.

If there is any urgency to this review process that I am not aware of, then please clarify and I'll elevate it to a higher priority.
