### 2025-07-17 d005ce0
Remove zero-electron barrier for gradients
