### 2025-07-17 f37c8ee
Update Windows runner version
