+++ .github/workflows/CI.yaml
@@ -347,8 +347,7 @@ jobs:
           path: build/mopac-*-mac.*
 
   windows-build:
-    # NOTE: windows-latest has been switched to Visual Studio 2022, which is not yet supported by Intel development tools
-    runs-on: windows-2019
+    runs-on: windows-latest
     timeout-minutes: 30
 
     steps:
