### 2025-07-17 8615c1a
Adjust Intel installer CLI options
