### 2025-08-27 b709f3f
Add QMMM support to MOZYME
