### 2025-08-27 f66535e
Replace C-C step function correction with continuous correction
