### 2025-08-28 0642674
Regularize C-C correction, add derivative, & move to compfg
