### 2025-08-28 283035f
Narrow correction region & increase polynomial degree of correction to 7
