+++ src/corrections/set_up_dentate.F90
@@ -220,14 +220,20 @@
   use molkst_C, only : numat, method_pm6, method_pm7, method_pm6_org, method_pm8
   implicit none
   integer :: i, j, k
-  double precision :: rab, rmin, rmax, sum
-    rmin = 1.26d0
-    rmax = 1.36d0
+  double precision :: rab, rmin, rmax, param1, param2, sum
+!
+! C-C double bond length = 1.34 Angstroms
+! C-C triple bond length = 1.20 Angstroms
+!
+    rmin = 1.21d0
+    rmax = 1.33d0
+    param1 = -5.d0
+    param2 = 25.d0
     if ( .not. (method_pm6 .or. method_pm7 .or. method_pm6_org .or. method_pm8)) then
       C_triple_bond_C = 0.d0
       return
     end if
-    sum = 0
+    sum = 0.d0
     do i = 1, numat
       if (nat(i) == 6) then
 !
@@ -238,14 +244,11 @@
           if (j > i) cycle
           if(nat(j) == 6) then
             rab = (coord(1,i) - coord(1,j))**2 + (coord(2,i) - coord(2,j))**2 + (coord(3,i) - coord(3,j))**2
-!
-! C-C double bond length = 1.34 Angstroms
-! C-C triple bond length = 1.20 Angstroms
-!
             if (rab < rmin**2) sum = sum + 1.d0
             if (rab >= rmin**2 .and. rab < rmax**2) then
               rab = (dsqrt(rab) - rmin) / (rmax - rmin)
-              sum = sum + 1.d0 - 10.d0*rab**3 + 15.d0*rab**4 - 6.d0*rab**5
+              sum = sum + 1.d0 - 10.d0*rab**3 + 15.d0*rab**4 - 6.d0*rab**5 + &
+              (param1+rab*param2)*(rab**3 - 3.d0*rab**4 + 3.d0*rab**5 - rab**6)
             end if
           end if
         end do
@@ -262,10 +265,12 @@
     use molkst_C, only : numat, method_pm6, method_pm7, method_pm6_org, method_pm8
     implicit none
     integer :: i, j, k
-    double precision :: rab, rmin, rmax, drab, dsum, dx, dy, dz
+    double precision :: rab, rmin, rmax, param1, param2, drab, dsum, dx, dy, dz
     double precision, intent (inout) ::  dxyz(3, numat)
-      rmin = 1.26d0
-      rmax = 1.36d0
+      rmin = 1.21d0
+      rmax = 1.33d0
+      param1 = -5.d0
+      param2 = 25.d0
       if ( .not. (method_pm6 .or. method_pm7 .or. method_pm6_org .or. method_pm8)) then
         return
       end if
@@ -282,7 +287,9 @@
                 dx = 2.d0*(coord(1,i) - coord(1,j))*drab
                 dy = 2.d0*(coord(2,i) - coord(2,j))*drab
                 dz = 2.d0*(coord(3,i) - coord(3,j))*drab
-                dsum = - 10.d0*3.d0*rab**2 + 15.d0*4.d0*rab**3 - 6.d0*5.d0*rab**4
+                dsum = - 10.d0*3.d0*rab**2 + 15.d0*4.d0*rab**3 - 6.d0*5.d0*rab**4 + &
+                (param1+rab*param2)*(3.d0*rab**2 - 3.d0*4.d0*rab**3 + 3.d0*5.d0*rab**4 - 6.d0*rab**5) + &
+                param2*(rab**3 - 3.d0*rab**4 + 3.d0*rab**5 - rab**6)
                 dxyz(1, i) = dxyz(1, i) + 12.d0*dsum*dx
[…]
