### 2025-08-30 251a752
Move population matrix from global pb to local popmat
