+++ src/forces/force.F90
@@ -42,7 +42,7 @@
       double precision :: time2, tscf, tder, time1, time3, a, b, c, &
         sum, const, summ, sum1, sym
       double precision, dimension(:), allocatable :: store, ff, oldf, &
-        velocity
+        velocity, popmat
       double precision, dimension (:,:), allocatable :: store_coord
       logical :: restrt, linear, debug, prnt, large, ts
       double precision, external :: ddot, dipole, reada, seconds
@@ -247,7 +247,11 @@
       '(10X," GRADIENT NORM SHOULD BE LESS THAN ABOUT",f4.1," FOR THERMO",&
       &/10X,'' TO GIVE ACCURATE RESULTS'')') sum
       end if
-      if ( .not. mozyme .and. .not. restrt) call mullik()
+      if ( .not. mozyme .and. .not. restrt) then
+        allocate(popmat(mpack))
+        call mullik(popmat)
+        deallocate(mullik(popmat))
+      end if
       if (tscf > 0.01D0) then
         write (iw, '(2/10X,''TIME FOR SCF CALCULATION ='',F8.2)') tscf
         write (iw, '( /10X,''TIME FOR DERIVATIVES     ='',F8.2)') tder
