### 2025-08-30 34617f0
Fix d orbitals in overlap matrix & add UHF case to mullik
