+++ src/properties/mullik.F90
@@ -20,7 +20,7 @@
 !
       use molkst_C, only : numat, nelecs, nclose, nopen, fract,  &
         keywrd, norbs, id, verson, method_pm6, uhf, nalpha, nbeta, &
-        numcal, escf, line, mpack
+        numcal, escf, line, mpack, nalpha_open, nbeta_open
 !
       use symmetry_C, only : jndex, namo
 !
@@ -29,7 +29,7 @@
       use common_arrays_C, only : nfirst, nlast, nat, coord, &
       & c, h, pb, tvec, eigs, q, eigb, p, ifact, cb
 !
-      use parameters_C, only : zs, zp, zd, betas, betap, tore
+      use parameters_C, only : zs, zp, zd, betas, betap, betad, tore, natorb
 !
       use chanel_C, only : igpt, gpt_fn, iw
 !
@@ -38,16 +38,14 @@
 !-----------------------------------------------
 !   L o c a l   V a r i a b l e s
 !-----------------------------------------------
-      integer :: i, if, il, im1, k, ii, j, jf, jl, ij, icalcn=0, mo_l, mo_u
+      integer :: i, if, il, im1, k, ii, j, jf, jl, ij, icalcn=0, mo_l, mo_u, norbi, norbj
       character, dimension(:), allocatable :: namo_tmp*4
       character :: paras*1
       double precision, dimension(norbs) :: eig
-      double precision, dimension(:), allocatable :: store, vecs, store_h
+      double precision, dimension(:), allocatable :: store, vecs, overlap
       double precision, dimension(:, :), allocatable :: work
-      double precision :: bi, bj, sum, summ, q2(numat)
+      double precision :: bi(9), bj(9), sum, summ, q2(numat)
       logical :: graph, graph_formatted, namo_ok
-! GBR_new_addition
-      integer :: nlower
 !
       save :: graph, graph_formatted, icalcn, mo_l, mo_u
 !
@@ -66,8 +64,7 @@
 !*********************************************************************
 
 
-     allocate(store((norbs*(norbs + 1))/2), store_h((norbs*(norbs + 1))/2), &
-       vecs(norbs**2), stat = i)
+     allocate(p_mullik(mpack), overlap(mpack), vecs(norbs**2), stat = i)
      if (i /= 0) then
         write(iw,*)" Unable to allocate memory for eigenvector matrices in MULLIK"
         call mopend("Unable to allocate memory for eigenvector matrices in MULLIK")
@@ -110,36 +107,46 @@
       do i = 1, numat
         q(i) = tore(nat(i) ) - q2(i)
       end do
-      store_h = h
+      ! overlap matrix construction from to_screen, including d orbitals
+      overlap = 0.d0
       do i = 1, numat
         if = nfirst(i)
-        il = nlast(i)
         im1 = i - 1
-        bi = betas(nat(i))
-        do k = if, il
-          ii = (k*(k - 1))/2
-          do j = 1, im1
-            jf = nfirst(j)
-            jl = nlast(j)
-            bj = betas(nat(j))
-!  THE  +1.D-14 IS TO PREVENT POSSIBLE ERRORS IN THE DIAGONALIZATION.
-            ij = ii + jf
-            h(ij) = 2.D0*h(ij)/(bi + bj) + 1.D-14
-            store(ij) = h(ij)
-            bj = betap(nat(j))
-            bj = betap(nat(j))
[…]
