### 2025-08-31 b17753f
Avoid use of mpack in mixed MOPAC/MOZYME code paths (mullik)
