### 2025-08-31 555d182
Add missing variable declarations in mullik
