### 2025-08-31 0e796b5
Synchronize density calculation in mullik w/ iter
