### 2025-09-04 #8025 @RMeli
Hi @berquist, @awvwgk, and @godotalgorithm, just checking in. Could you please shortly state how this review is progressing?
