### 2025-09-10 0d2819b
Check (relative) time limit in iter
