### 2025-09-11 41f09f2
Switch to powershell to get # of cores in Windows CI
