### 2025-09-11 a02182e
Fix powershell command
