### 2025-09-11 0d7d550
Remove last defunct references to Java/JMol
