### 2025-09-12 0e3f27c
Fix loop bounds when checking for lattice vectors in API
