### 2025-09-18 3237f82
Distinguish between PRECISE and PRECISION keywords
