### 2025-10-13 d21ff62
Set no-convergence error flags for iter timeout
