### 2025-10-13 7959ffc
Add more suggestions to the SCF failure message
