+++ src/output/writmo.F90
@@ -152,6 +152,13 @@
         write (iw, '(A)') ' '
         write (iw, '(A)') &
           ' CHECK YOUR GEOMETRY AND ALSO TRY USING SHIFT OR PULAY. '
+        write (iw, '(A)') ' '
+        write (iw, '(A)') &
+          ' CLOSED-SHELL ORGANIC MOLECULES SHOULD HAVE GOOD SCF CONVERGENCE, '
+        write (iw, '(A)') &
+          ' CHECK THEIR TOTAL CHARGE WITH THE ''LEWIS'' KEYWORD '
+        write (iw, '(A)') &
+          ' AND CHECK THEIR PROTONATION WITH THE ''ADD-H'' KEYWORD. '
         call geout (1)
         call mopend ('THE SCF CALCULATION FAILED.')
         return
