### 2025-10-13 f990056
Adjust format of error message
