+++ src/output/writmo.F90
@@ -156,9 +156,10 @@
         write (iw, '(A)') &
           ' CLOSED-SHELL ORGANIC MOLECULES SHOULD HAVE GOOD SCF CONVERGENCE, '
         write (iw, '(A)') &
-          ' CHECK THEIR TOTAL CHARGE WITH THE ''LEWIS'' KEYWORD '
+          ' CHECK THEIR TOTAL CHARGE WITH THE ''LEWIS'' KEYWORD, '
         write (iw, '(A)') &
           ' AND CHECK THEIR PROTONATION WITH THE ''ADD-H'' KEYWORD. '
+        write (iw, '(A)') ' '
         call geout (1)
         call mopend ('THE SCF CALCULATION FAILED.')
         return
