### 2025-10-24 b69ef95
Add API test to detect charge & PM6-D3H4 bugs
