### 2025-10-24 11c5bfa
Fix API PM6-D3H4 bug
