### 2025-10-24 754a1db
Fix API dipole with frozen atoms
