### 2025-10-28 ff49c2c
Set RHF & UHF flags in API
