### 2025-10-28 97055cd
Add RHF/UHF switch test for API
