### 2025-10-28 ee30135
Fix RHF/UHF test reference data & remove case w/ SCF hysteresis
