### 2025-12-03 2b24cef
Add examples to repo and packaging
