### 2025-12-04 #8025 @godotalgorithm
The new website is now live on [GitHub Pages](https://openmopac.github.io). I'll be meeting with the original MOPAC developer next week about decommissioning the old website, and I'll begin serving this website (and all websites under the MOPAC GitHub organization) from the openmopac.net domain once I have his blessing.

I'll address the remaining referee GitHub Issues over the next few days, and we can finally continue with the review.
