### 2025-12-04 openmopac/mopac#272 @godotalgorithm
There are now a few basic examples in the `/examples` directory with reference outputs, which will be packaged with future releases of MOPAC.
