### 2025-12-04 openmopac/mopac#264 @godotalgorithm
The new website is now live on [GitHub Pages](https://openmopac.github.io/). This is a temporary location until I meet with Jimmy Stewart next week to finalize the website. After I get his approval, I will set GitHub Pages to serve all MOPAC websites from openmopac.net, which will retire the old website.

The new website retains the old manual webpages, and there are still a few references to MOPAC2016 there. However, they are much less prominent, and I will be cleaning them up eventually now that the whole website is under version control.
