+++ README.md
@@ -15,109 +15,44 @@ Commercial versions of MOPAC are no longer supported, and all MOPAC users are en
 
 MOPAC is actively maintained and curated by the [Molecular Sciences Software Institute (MolSSI)](https://molssi.org).
 
-## Installation
+For detailed information about MOPAC, see its [website](https://openmopac.github.io). A brief summary for developers is provided below.
 
-Open-source MOPAC is available through multiple distributon channels, and it can also be compiled from source using CMake.
-In addition to continuing the distribution of self-contained installers on the
-[old commercial website](http://openmopac.net/Download_MOPAC_Executable_Step2.html) and here on GitHub,
-MOPAC can also be installed using multiple package managers and accessed through containers.
+## About
 
-### Self-contained installers
+MOPAC is a popular semiempirical quantum chemistry package that was first released in 1983.
+MOPAC can perform quantum mechanical calculations of molecules and materials, much like *ab initio* quantum chemistry packages such as [Gaussian](https://www.gaussian.com).
+Semiempirical calculations are around 1000x faster than *ab initio* calculations, but they are usually less accurate because they rely on
+simplified, minimal-basis model Hamiltonians that are fit to experimental data rather than more predictive theories.
+The speed and ease-of-use of MOPAC makes it a good teaching tool, either to teach physical chemistry concepts to students or to prepare yourself
+to use more expensive quantum chemistry software. It can also be useful for high-throughput calculations in screening and informatics applications.
 
-Self-contained graphical installers for Linux, Mac, and Windows are available on GitHub for each release,
-which are constructed using the [Qt Installer Framework](https://doc.qt.io/qtinstallerframework/).
+## Building
 
-While the installers are meant to be run from a desktop environment by default, they can also be run from a command line without user input.
-On Linux, the basic command-line installation syntax is:
+Pre-built versions of MOPAC on Linux, Mac, and Windows are available with every [GitHub Release](https://github.com/openmopac/mopac/releases)
+and also available on the Conda package manager using the command:
+```
+conda install -c conda-forge mopac
+```
 
-`./mopac-x.y.z-linux.run install --accept-licenses --confirm-command --root type_installation_directory_here`
-
-For more information on command-line installation, see the [Qt Installer Framework Documentation](https://doc.qt.io/qtinstallerframework/ifw-cli.html).
-
-Linux installations without a desktop environment may not have the shared libraries required for the graphical installers,
-and there have also been isolated reports of problems with the Qt installer on other platforms. A minimal, compressed-archive installer
-is available for each platform as an alternative for users that have problems with the Qt installer.
-
-The minimum glibc version required for the precompiled version of MOPAC on Linux is currently 2.17.
-
-#### Library path issues
-
-The pre-built MOPAC executables use the RPATH system on Mac and Linux to connect with its shared libraries,
-including the `libiomp5` Intel OpenMP redistributable library. The `libiomp5` library is not properly versioned, and the recent version used by
-MOPAC is not compatible with older versions that might also exist on a user's machine. If a directory containing an old version of `libiomp5`
-is in the shared library path (`LD_LIBRARY_PATH` on Linux, `DYLD_LIBRARY_PATH` on Mac), this will override the RPATH system, link MOPAC to the
-wrong library, and cause an error in MOPAC execution. On Mac, this can be fixed by switching the offending directories to the failsafe shared library
-path, `DYLD_FALLBACK_LIBRARY_PATH`. On Linux, the use of `LD_LIBRARY_PATH` is generally discouraged for widespread use, and there is no simple
[…]
