### 2025-12-04 openmopac/mopac#265 @godotalgorithm
The MOPAC installation information is now mostly synchronized except for some stray information in the manual that I will clean up after the final migration step next week.

The Sphinx-based website is now decommissioned, and the old openmopac.net website should be replaced by the new website on GitHub Pages next week. The primary download & installation information is now on the [download page](https://openmopac.github.io/download/), and the repo README file contains a consistent, limited summary of this information.

I will hold off on closing this Issue until the old website has been decommission next week.
