### 2025-12-04 bbe5964
Update README.md (clean-up & streamlining)
