### 2025-12-04 a7bd24b
Update /examples/README.md w/ further info
