+++ examples/README.md
@@ -43,3 +43,10 @@ gfortran example.F90 ../include/mopac_wrapper.F90 ../include/mopac_wrapper_inter
 Additional linking options such as `-Wl,-rpath=../lib` may be necessary if the MOPAC shared library is not in your path.
 `example.F90_output` is a reference copy of the standard output generated by this example.
 
+## Further information
+
+MOPAC's features and command-line functionality are primarily determined by keywords at the top of the input file.
+Online [guides](https://openmopac.github.io/guides/) provide instructions on how to use the most
+popular MOPAC keywords and features, and the online [manual](https://openmopac.github.io/Manual/) has a more detailed and
+comprehensive description of MOPAC's keywords and features. Further information about the API is available in the
+[API guide](https://openmopac.github.io/guides/api/).
