### 2025-12-04 508961f
Update /examples/README.md (typo)
