### 2025-12-10 openmopac/mopac#265 @godotalgorithm
Ok, MOPAC's web presence is fully cleaned up now, so this issue should be resolved.
