### 2026-02-04 8ade762
Update reference to SQM perspective paper
