+++ paper.bib
@@ -1,12 +1,25 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% https://bibdesk.sourceforge.io/
 
-%% Created for Jonathan Moussa at 2025-05-23 12:55:02 -0400 
+%% Created for Jonathan Moussa at 2026-02-04 11:58:41 -0500 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
 
 
+
+@article{SQM_perspective,
+	annote = {doi: 10.1021/acs.jpca.5c03425},
+	author = {Moussa, Jonathan E. },
+	date-added = {2026-02-04 11:55:58 -0500},
+	date-modified = {2026-02-04 11:58:31 -0500},
+	doi = {10.1021/acs.jpca.5c03425},
+	journal = {The Journal of Physical Chemistry A},
+	pages = {8465--8477},
+	title = {The Enduring Relevance of Semiempirical Quantum Mechanics},
+	volume = {129},
+	year = {2025},
+	bdsk-url-1 = {https://doi.org/10.1021/acs.jpca.5c03425}}
 
 @misc{Gaussian,
 	address = {Wallingford, CT %! Gaussian 16},
@@ -16,18 +29,6 @@
 	date-modified = {2025-05-23 12:54:57 -0400},
 	title = {Gaussian 16 Rev. C.01},
 	year = {2016}}
-
-@misc{SQM_perspective,
-	archiveprefix = {arXiv},
-	author = {Jonathan E. Moussa},
-	date-added = {2025-05-23 12:46:55 -0400},
-	date-modified = {2025-05-23 12:47:11 -0400},
-	eprint = {2505.13424},
-	primaryclass = {physics.chem-ph},
-	title = {The Enduring Relevance of Semiempirical Quantum Mechanics},
-	url = {https://arxiv.org/abs/2505.13424},
-	year = {2025},
-	bdsk-url-1 = {https://arxiv.org/abs/2505.13424}}
 
 @article{ULYSSES,
 	annote = {doi: 10.1021/acs.jcim.2c00757},
@@ -328,4 +329,5 @@
 	publisher = {GitHub},
 	title = {mopactools},
 	url = {https://github.com/openmopac/mopactools},
-	year = {2025}}
+	year = {2025},
+	bdsk-url-1 = {https://github.com/openmopac/mopactools}}
