### 2026-02-04 #8025 @godotalgorithm
I updated the reference to the published version, as requested.

I would like to release another patch version of MOPAC prior to publication, just to include some of the minor adjustments that occurred because of this review process (e.g. the addition of examples to the repo & distribution). However, I'm wondering if I could also get a DOI for this upcoming JOSS paper so that I can adjust the recommended citation in the MOPAC version associated with the paper. Otherwise, I'm going to release yet another patch immediately after the release of this paper to update the recommended citation. Either case is fine, it would just be a bit more streamlined if they could be synchronized. I appreciate that an early DOI assignment may be incompatible with JOSS's publication process.
