### 2026-02-05 #8025 @RMeli
> I would like to release another patch version of MOPAC prior to publication, just to include some of the minor adjustments that occurred because of this review process

Yes, that would be great.

> However, I'm wondering if I could also get a DOI for this upcoming JOSS paper so that I can adjust the recommended citation in the MOPAC version associated with the paper.

DOIs are generated only on publication, but looks like the process is predictable and the numbering is associated to the PR issue number. Therefore, the DOI here should become

```
https://doi.org/10.21105/joss.08025
```
