### 2026-02-16 #8025 @godotalgorithm
I've made the patch release for the JOSS paper. I decided against an early citation, since I don't have a volume/issue/page number in addition to the DOI. I'll just release another patch once the JOSS paper is out to update the citation information.

Please tell me if there is anything else I need to do to wrap up the publication process.
