### 2026-02-17 #8025 @RMeli
@godotalgorithm one last thing, then we are good to go: can you please change the Zenodo metadata to match the record with the title of the paper? If possible, you should also add your ORCID on the Zenodo profile.

> Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
