### 2026-02-17 #8025 @godotalgorithm
Ok, I adjusted that Zenodo record name to the title of the JOSS paper and added my ORCID number to that record. Do those changes satisfy your request? I think that I interpreted what you said correctly, but I'm not 100% sure.
