### 2026-02-17 #8025 @RMeli
Perfect @godotalgorithm, everything LGTM.
