### 2026-02-18 #8025 @godotalgorithm
I would also like to thank @berquist, @awvwgk, and @RMeli for your efforts and patience. This was my first fully open paper review, and it was a very positive experience overall. It ended up being a bit slow, but that was mostly my fault, and I'm grateful for the accommodation.
