### 2026-02-20 #8025 @rkurchin
Hi @godotalgorithm, track AEiC here – I'm not certain that this is what is causing the issue above, but I do see in your `.bib` file that you have an ISBN for ULYSSES that is indeed 9 characters long; since that's a journal paper anyway, that field shouldn't be required, so could you just remove that field from that entry and we can try processing this again?
