### 2026-02-21 55a102e
Clean stray fields from .bib file
