+++ paper.bib
@@ -1,7 +1,7 @@
 %% This BibTeX bibliography file was created using BibDesk.
 %% https://bibdesk.sourceforge.io/
 
-%% Created for Jonathan Moussa at 2026-02-04 11:58:41 -0500 
+%% Created for Jonathan Moussa at 2026-02-21 11:05:16 -0500 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
@@ -10,7 +10,7 @@
 
 @article{SQM_perspective,
 	annote = {doi: 10.1021/acs.jpca.5c03425},
-	author = {Moussa, Jonathan E. },
+	author = {Moussa, Jonathan E.},
 	date-added = {2026-02-04 11:55:58 -0500},
 	date-modified = {2026-02-04 11:58:31 -0500},
 	doi = {10.1021/acs.jpca.5c03425},
@@ -33,40 +33,27 @@
 @article{ULYSSES,
 	annote = {doi: 10.1021/acs.jcim.2c00757},
 	author = {Menezes, Filipe and Popowicz, Grzegorz M.},
-	date = {2022/08/22},
 	date-added = {2025-05-23 12:46:13 -0400},
-	date-modified = {2025-05-23 12:46:29 -0400},
+	date-modified = {2026-02-21 11:05:04 -0500},
 	doi = {10.1021/acs.jcim.2c00757},
-	isbn = {1549-9596},
 	journal = {Journal of Chemical Information and Modeling},
-	journal1 = {Journal of Chemical Information and Modeling},
-	journal2 = {J. Chem. Inf. Model.},
-	month = {08},
 	number = {16},
 	pages = {3685--3694},
-	publisher = {American Chemical Society},
 	title = {ULYSSES: An Efficient and Easy to Use Semiempirical Library for C++},
-	type = {doi: 10.1021/acs.jcim.2c00757},
-	url = {https://doi.org/10.1021/acs.jcim.2c00757},
 	volume = {62},
 	year = {2022},
-	year1 = {2022},
 	bdsk-url-1 = {https://doi.org/10.1021/acs.jcim.2c00757}}
 
 @article{Sparrow,
 	abstract = {Semi-empirical quantum chemical approaches are known to compromise accuracy for the feasibility of calculations on huge molecules. However, the need for ultrafast calculations in interactive quantum mechanical studies, high-throughput virtual screening, and data-driven machine learning has shifted the emphasis toward calculation runtimes recently. This comes with new constraints for the software implementation as many fast calculations would suffer from a large overhead of the manual setup and other procedures that are comparatively fast when studying a single molecular structure, but which become prohibitively slow for high-throughput demands. In this work, we discuss the effect of various well-established semi-empirical approximations on calculation speed and relate this to data transfer rates from the raw-data source computer to the results of the visualization front end. For the former, we consider desktop computers, local high performance computing, and remote cloud services in order to elucidate the effect on interactive calculations, for web and cloud interfaces in local applications, and in world-wide interactive virtual sessions. The models discussed in this work have been implemented into our open-source software SCINE Sparrow.},
 	author = {Bosia, Francesco and Zheng, Peikun and Vaucher, Alain and Weymuth, Thomas and Dral, Pavlo O. and Reiher, Markus},
 	date-added = {2025-05-23 12:44:59 -0400},
-	date-modified = {2025-05-23 12:45:07 -0400},
+	date-modified = {2026-02-21 11:04:32 -0500},
 	doi = {10.1063/5.0136404},
-	eprint = {https://pubs.aip.org/aip/jcp/article-pdf/doi/10.1063/5.0136404/16703358/054118\_1\_online.pdf},
-	issn = {0021-9606},
 	journal = {The Journal of Chemical Physics},
[…]
