### 2026-02-21 #8025 @godotalgorithm
@rkurchin I cleaned out the unnecessary fields from the `.bib` file. Hopefully it works now.
