### 2026-02-23 #8025 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
