### 2026-02-26 #8025 @godotalgorithm
If you are waiting on me for approval of the proofs, it looks good to me, and I just got the thumbs up from my co-author as well.
