### 2026-03-06 #8025 @RMeli
@godotalgorithm we are now waiting for the EiC to have a final look at the submission and accept it for publication. EiCs are quite busy at the moment due to the large number of submission, sorry for the additional delay.
