### 2026-03-15 #8025 @Kevin-Mattheus-Moerman
@godotalgorithm As AEiC for JOSS I will now help to process this submission for acceptance in JOSS (and apologies for the delay in getting this finalization process started!). Below are some final checks:

- [x] Project features contributing guidelines as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which may require your attention:

## Checks on repository👋 Hi, congrats on being nearly done! As AEiC for JOSS I will now help to process this submission for acceptance in JOSS (and apologies for the delay in getting this finalization process started!). Below are some final checks:

- [x] Copyedit pass/checked for typos
